third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/getProperties-expected.txt

This is a testharness.js-based test.
PASS Calling StylePropertyMap.getProperties on an empty property model returns a zero-length array
FAIL StylePropertyMap.getProperties returns properties in sorted order assert_array_equals: property 0, expected "color" but got "--A"
Harness: the test ran to completion.
